fix(multi_agents): rename mas->multi_agents
