--- ai_platform_engineering/agents/argocd/MAINTAINERS.md
@@ -1,6 +1,0 @@
-# Maintainers
-
-- [sriaradhyula](https://github.com/sriaradhyula)
-- [sbraicu](https://github.com/sbraicu)
-- [artdroz](https://github.com/artdroz)
-- [rehanthestar21](https://github.com/rehanthestar21)
