--- ai_platform_engineering/agents/argocd/SECURITY.md
@@ -1,24 +1,0 @@
-## Report a security issue
-
-The CNOE project team welcomes security reports and is committed to
-providing prompt attention to security issues. Security issues should be
-reported privately via [cnoe-steering@googlegroups.com](mailto:cnoe-steering@googlegroups.com).
-Security issues should not be reported via the public GitHub Issue tracker.
-
-## Vulnerability coordination
-
-Remediation of security vulnerabilities is prioritized by the project team. The
-project team coordinates remediation with third-party project stakeholders via
-[GitHub Security Advisories](https://help.github.com/en/github/managing-security-vulnerabilities/about-github-security-advisories).
-Third-party stakeholders may include the reporter of the issue, affected direct or indirect
-users of CNOE, and maintainers of upstream dependencies if applicable.
-
-Downstream project maintainers and CNOE users can request participation in
-coordination of applicable security issues by sending your contact email address,
-GitHub username(s) and any other salient information to [cnoe-steering@googlegroups.com](mailto:cnoe-steering@googlegroups.com).
-Participation in security issue coordination processes is at the discretion of the CNOE team.
-
-## Security advisories
-
-The project team is committed to transparency in the security issue disclosure
-process. The CNOE team announces security issues via project GitHub Release notes.
