+++ ai_platform_engineering/agents/argocd/a2a_agent_client/agent.py
@@ -12,7 +12,7 @@ from ai_platform_engineering.agents.argocd.a2a_agent_client.agentcard import (
 from ai_platform_engineering.utils.a2a.a2a_remote_agent_connect import (
     A2ARemoteAgentConnectTool,
 )
-from ai_platform_engineering.mas.platform_engineer.prompts import get_agent_system_prompt
+from ai_platform_engineering.multi_agents.platform_engineer.prompts import get_agent_system_prompt
 
 model = LLMFactory().get_llm()
