+++ ai_platform_engineering/graph.py
@@ -1,3 +1,3 @@
-from ai_platform_engineering.mas.platform_engineer.supervisor_agent import AIPlatformEngineerMAS
+from ai_platform_engineering.multi_agents.platform_engineer.supervisor_agent import AIPlatformEngineerMAS
 
 graph = AIPlatformEngineerMAS().get_graph()
