+++ ai_platform_engineering/multi_agents/platform_engineer/protocol_bindings/fastapi/main.py
@@ -6,7 +6,7 @@ import logging
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 
-from ai_platform_engineering.mas.platform_engineer.supervisor_agent import AIPlatformEngineerMAS
+from ai_platform_engineering.multi_agents.platform_engineer.supervisor_agent import AIPlatformEngineerMAS
 from ai_platform_engineering.utils.models.generic_agent import UserPrompt
 
 logging.basicConfig(
